backend/requirements.txt
Django>=5.0.6,<5.1
psycopg2==2.9.9,<2.10
uWSGI>=2.0.25.1,<2.1
djangorestframework>=3.15.1,<3.16
markdown>=3.6,<3.7
django-filter>=24.2,<25.0
dj-rest-auth>=6.0.0,<6.1
djangorestframework-simplejwt>=5.3.1,<5.4
django-allauth>=0.62.0,<0.63
requests